change ctx manager to fixture
