+++ tests/unit/core/exporter/exportable_code/demo/demo_package/streamer/test_streamer.py
@@ -4,15 +4,11 @@
 
 import tempfile
 from pathlib import Path
+from typing import Iterator
 
+import cv2
+import numpy as np
 import pytest
-
-from tests.test_helpers import (
-    generate_random_image_folder,
-    generate_random_single_image,
-    generate_random_single_video,
-    generate_random_video_folder,
-)
 
 CameraStreamer = None
 DirStreamer = None
@@ -50,13 +46,116 @@ def fxt_import_module():
     get_streamer = func1
 
 
+@pytest.fixture()
+def random_image_folder(tmp_path, width: int = 480, height: int = 360, number_of_images: int = 10) -> str:
+    """
+    Generates a folder with random images, cleans up automatically if used in a `with` statement
+
+    Parameters:
+        width (int): height of the images. Defaults to 480.
+        height (int): width of the images. Defaults to 360.
+        number_of_images (int): number of generated images. Defaults to 10.
+
+    Returns:
+        Iterator[str]: The temporary directory
+    """
+    for n in range(number_of_images):
+        temp_file = str(tmp_path / f"{n}.jpg")
+        _write_random_image(width, height, temp_file)
+
+    return str(tmp_path)
+
+
+@pytest.fixture()
+def random_video_folder(
+    tmp_path,
+    width: int = 480,
+    height: int = 360,
+    number_of_videos: int = 10,
+    number_of_frames: int = 150,
+) -> Iterator[str]:
+    """
+    Generates a folder with random videos, cleans up automatically if used in a `with` statement
+
+    Parameters:
+        width (int): Width of the video. Defaults to 480.
+        height (int): Height of the video. Defaults to 360.
+        number_of_videos (int): Number of videos to generate. Defaults to 10.
+        number_of_frames (int): Number of frames in each video. Defaults to 150.
+
+    Returns:
+        Iterator[str]: A temporary directory with videos
+    """
+
+    for n in range(number_of_videos):
+        temp_file = str(tmp_path / f"{n}.mp4")
+        _write_random_video(width, height, number_of_frames, temp_file)
+
+    return str(tmp_path)
+
+
+@pytest.fixture()
+def random_single_image(tmp_path, width: int = 480, height: int = 360) -> str:
+    """
+    Generates a random image, cleans up automatically if used in a `with` statement
+
+    Parameters:
+        width (int): Width of the image. Defaults to 480.
+        height (int): Height of the image. Defaults to 360.
+
+    Returns:
+        Iterator[str]: Path to an image file
+    """
+
+    temp_file = str(tmp_path / "temp_image.jpg")
+    _write_random_image(width, height, temp_file)
+
+    return temp_file
+
+
+@pytest.fixture()
+def random_single_video(tmp_path, width: int = 480, height: int = 360, number_of_frames: int = 150) -> str:
+    """
+    Generates a random video, cleans up automatically if used in a `with` statement
+
+    Parameters:
+        width (int): Width of the video. Defaults to 480.
+        height (int): Height of the video. Defaults to 360.
+        number_of_frames (int): Number of frames in the video. Defaults to 150.
+
+    Returns:
+        Iterator[str]: Path to a video file
+    """
+    temp_file = str(tmp_path / "temp_video.mp4")
+    _write_random_video(width, height, number_of_frames, temp_file)
+
+    return temp_file
+
+
+def _write_random_image(width: int, height: int, filename: str) -> None:
+    img = np.uint8(np.random.random((height, width, 3)) * 255)
+    cv2.imwrite(filename, img)
+
+
+def _write_random_video(width: int, height: int, number_of_frames: int, filename: str) -> None:
+    fourcc = cv2.VideoWriter_fourcc(*"mp4v")
+    f = filename
+    videowriter = cv2.VideoWriter(f, fourcc, 30, (width, height))
+
+    for _ in range(number_of_frames):
+        img = np.uint8(np.random.random((height, width, 3)) * 255)
+        videowriter.write(img)
+
+    videowriter.release()
+
+
 class TestStreamer:
     @staticmethod
     def assert_streamer_element(streamer) -> None:
         for element in streamer:
             assert element.shape == (360, 480, 3)
 
-    def test_image_streamer_with_single_image(self):
+    def test_image_streamer_with_single_image(self, random_single_image):
         """
         <b>Description:</b>
         Test that ImageStreamer works correctly with a single image as input
@@ -71,11 +170,10 @@ class TestStreamer:
         1. Create ImageStreamer
         2. Request image from streamer
         """
-        with generate_random_single_image(height=360, width=480) as path:
-            streamer = ImageStreamer(path)
-            self.assert_streamer_element(streamer)
+        streamer = ImageStreamer(random_single_image)
+        self.assert_streamer_element(streamer)
 
-    def test_dir_streamer_with_folder(self):
+    def test_dir_streamer_with_folder(self, random_image_folder):
         """
         <b>Description:</b>
         Test that DirStreamer works correctly with a folder of images as input
@@ -90,11 +188,10 @@ class TestStreamer:
         1. Create DirStreamer
         2. Request images from streamer
         """
-        with generate_random_image_folder(height=360, width=480) as path:
-            streamer = DirStreamer(path)
-            self.assert_streamer_element(streamer)
+        streamer = DirStreamer(random_image_folder)
+        self.assert_streamer_element(streamer)
 
-    def test_video_streamer_with_single_video(self):
+    def test_video_streamer_with_single_video(self, random_single_video):
         """
         <b>Description:</b>
         Test that VideoStreamer works correctly with a single video as input
@@ -109,11 +206,10 @@ class TestStreamer:
         1. Create VideoStreamer
         2. Request frame from VideoStreamer
         """
-        with generate_random_single_video(height=360, width=480) as path:
-            streamer = VideoStreamer(path)
-            self.assert_streamer_element(streamer)
+        streamer = VideoStreamer(random_single_video)
+        self.assert_streamer_element(streamer)
 
-    def test_video_streamer_with_loop_flag(self):
+    def test_video_streamer_with_loop_flag(self, random_single_video):
         """
         <b>Description:</b>
         Test that VideoStreamer works correctly with a loop flag
@@ -129,15 +225,14 @@ class TestStreamer:
         1. Create VideoStreamer
         2. Request frames from streamer
         """
-        with generate_random_single_video(height=360, width=480, number_of_frames=100) as path:
-            streamer = VideoStreamer(path, loop=True)
+        streamer = VideoStreamer(random_single_video, loop=True)
 
-            for index, frame in enumerate(streamer):
-                assert frame.shape[-1] == 3
-                if index > 200:
-                    break
+        for index, frame in enumerate(streamer):
+            assert frame.shape[-1] == 3
+            if index > 200:
+                break
 
-    def test_video_streamer_with_single_image(self):
+    def test_video_streamer_with_single_image(self, random_single_video):
         """
         <b>Description:</b>
         Test that VideoStreamer works correctly with a single image as input
@@ -152,11 +247,10 @@ class TestStreamer:
         1. Create VideoStreamer
         2. Request frame from VideoStreamer
         """
-        with generate_random_single_video(height=360, width=480) as path:
-            streamer = VideoStreamer(path)
-            self.assert_streamer_element(streamer)
+        streamer = VideoStreamer(random_single_video)
+        self.assert_streamer_element(streamer)
 
-    def test_invalid_inputs_to_get_streamer(self):
+    def test_invalid_inputs_to_get_streamer(self, random_video_folder):
         """
         <b>Description:</b>
         Test that get_streamer does not allow invalid inputs
@@ -187,13 +281,18 @@ class TestStreamer:
         with tempfile.TemporaryDirectory() as empty_dir, pytest.raises(Exception, match="Can't find the camera"):
             get_streamer(empty_dir)
 
-        with generate_random_video_folder() as path, pytest.raises(Exception, match="Can't find the camera"):
-            get_streamer(path)
+        with pytest.raises(Exception, match="Can't find the camera"):
+            get_streamer(random_video_folder)
 
         with pytest.raises(Exception, match="Can't find the camera"):
             get_streamer("not_a_file")
 
-    def test_valid_inputs_to_get_streamer(self):
+    def test_valid_inputs_to_get_streamer(
+        self,
+        random_single_video,
+        random_single_image,
+        random_image_folder,
+    ):
         """
         <b>Description:</b>
         Test that get_streamer return the correct Streamer class for each input
@@ -213,17 +312,14 @@ class TestStreamer:
         4. Call get_streamer with a camera index
         5. Call get_streamer with the threaded argument
         """
-        with generate_random_single_video() as path:
-            streamer = get_streamer(path)
-            assert isinstance(streamer, VideoStreamer)
+        streamer = get_streamer(random_single_video)
+        assert isinstance(streamer, VideoStreamer)
 
-        with generate_random_single_image() as path:
-            streamer = get_streamer(path)
-            assert isinstance(streamer, ImageStreamer)
+        streamer = get_streamer(random_single_image)
+        assert isinstance(streamer, ImageStreamer)
 
-        with generate_random_image_folder() as path:
-            streamer = get_streamer(path)
-            assert isinstance(streamer, DirStreamer)
+        streamer = get_streamer(random_image_folder)
+        assert isinstance(streamer, DirStreamer)
 
         streamer = get_streamer("0")
         assert isinstance(streamer, CameraStreamer)
@@ -231,7 +327,7 @@ class TestStreamer:
         streamer = get_streamer(input_stream="0", threaded=True)
         assert isinstance(streamer, ThreadedStreamer)
 
-    def test_video_file_fails_on_image_streamer(self):
+    def test_video_file_fails_on_image_streamer(self, random_single_video):
         """
         <b>Description:</b>
         Test that ImageStreamer raises an exception if a video is passed
@@ -245,8 +341,8 @@ class TestStreamer:
         <b>Steps</b>
         1. Attempt to create ImageStreamer
         """
-        with generate_random_single_video() as path, pytest.raises(RuntimeError):
-            ImageStreamer(path)
+        with pytest.raises(RuntimeError):
+            ImageStreamer(random_single_video)
 
     def test_camera_streamer(self):
         """
